Rename misnamed Apply page folder
We move 'example-page' to 'example-section'
