+++ server/form-pages/apply/index.ts
@@ -2,7 +2,7 @@
 import BaseForm from '../baseForm'
 import { Form } from '../utils/decorators'
 import CheckYourAnswers from './check-your-answers'
-import ExampleSection from './example-page'
+import ExampleSection from './example-section'
 
 @Form({ sections: [ExampleSection, CheckYourAnswers] })
 export default class Apply extends BaseForm {}
